remove check_status option for cache additions
This would cause 2 netlink connections to be made. For each packet.

I don't think it was even necessary; there would be notifications about blocked or excepted packets anyway, and the cache can be updated then. Initial tests seem too confirm this. If we do end up needing a status check, we should only do it if new information is added