Refactor handle_dns_{answer,query}() code; move to lib/
In the future, we want to be able to call the functions that analyze
DNS packets (handle_dns_query() and handle_dns_answer()) from other
places than spind so we need to move the functions from spind/ to lib/.
Basically, the only result of calling those functions will now be that
they yield one or more dns_pkt_info_t structures.

Any interaction with the node_cache and dns_cache is not the
responsibility of handle_dns_query() and handle_dns_answer() since
updating those caches is spind-specific and doesn't have anything to do
with analyzing the DNS packets. Instead, handle_dns_query() and
handle_dns_answer() now call a user-defined callback function to
handle any dns_pkt_info_t structures that are produced by the
handle_dns_query() and handle_dns_answer() functions. This mechanism
is now used in spind/core2nflog_dns.c to add entries to the dns_cache
and node_cache and to send a MQTT command.