Large refactor of iplist functionality
The three internal ip lists (ignore, block, allow), had quite a few interdependencies in the code.

We are removing SPIN/commands as much as possible (hopefully reduce it to broadcast information only in the future, perhaps rename it then), and replace such calls by RPC calls.

The IP lists are tied to the node cache (for efficiency), the firewall rules, and pubsub (to publish the full lists when they change). All actual updating from the client(s) should be done by RPC calls.

This patch moves things around and adds those calls in the RPC handler.