Remove all leftover code pertaining to mqtt topic SPIN/commands
Now that we have implemented all functionality in RPC commands, the handling of SPIN/commands can be removed. This patch removes the rest of the handler code, as well as the subscription code and the configuration data for SPIN/commands. The only MQTT channel spind now uses is SPIN/traffic.