Introduce a passive mode for spind, both run-time and compile-time
The run-time option is useful if you want to run spind without having
spind interfere with the network traffic on your machine. It allows
spind to be started as a non-root user. This may be useful in
combination with the extsrc facility.

The "active" mode (i.e. the way we always used spind up till now) relies
on the Netlink and Netfilter APIs (only available on Linux, as far as I
know). The new compile-time option allows you to compile spind on a
system which does not support those APIs (although with limited
functionality).

Not perfect yet; for instance, we still invoke iptables every now and
then.

Maybe there's an even better name for this feature.