Move cmp_flow_keys() from spinhook.c to node_cache.c
Not sure that is the right place but I think it's better than
spinhook.c. We could consider making the function static (it's only
used in node_cache.c) or we can move it to util.c, where other cmp_*()
functions live but then we also need the dev_flow_key_t type.