Use va_list functions correctly
Call va_end() only after we're done with the va_list, and if we want
to reuse a va_list, we have to call va_start() again. We could also
rewrite that piece of code to build a string once and use that twice
but I'm not doing that right now.