check_for_existing_node_with_mac(): don't merge identical nodes
No need to merge the nodes if the two nodes are the same. This can
happen in the following case: spinhook_traffic() first checks whether
either of the two nodes is a device or not. If none of them are a
device, node_cache_update_arp() is called. This function can result in
a MAC address being set for a node. As a result, it is possible that
the call to node_cache_find_by_mac() in
check_for_existing_node_with_mac() returns a node that is equal to the
"node" variable. Of course, we should not merge them.

This problem could be solved in a number of different ways, e.g.
by changing merge_nodes() to do nothing when two identical nodes
are passed. Not sure what is the best way to do it.

Found using Smart_Light-Philips-Hue_1.pcap.