spin-pcap-reader: aggregate pkt_infos and send them once a second
With "busy" PCAPs, this should reduce the traffic between the
PCAP reader and the SPIN daemon quite significantly in most cases.

This was actually really straightforward to implement given the fact
that the flow_list_*() functions were already available. Code in
send_flows() more or less based on flow_list2json() in spindata.c.