Use a separate config setting for mosquitto password file and spinweb password file
spinweb_password_file enables HTTP authentication on the spinweb pages and API
spinweb_pubsub_password_file enables MQTT user authentication on the mosquitto instance (if run by spin)

pubsub is more important, as this gives a user access to traffic data; but enableing http authentication on spinweb can provide an additional layer of protection. Note that when both are set, the user has to authenticate twice, as the spinweb client cannot pass on authentication information.