Handle mosquitto as a subprocess better with additional config
In some circumstances, such as shutting down spind with a signal 15, mosquitto (when started by spin) was not shut down correctly. By adding optional configuration to have mosquitto store its pid in a file, this can be handled better (for instance, in packaging where spin is started and stopped by an init.d script).