Update vis.js to 7.4.9
This fixes the issue where the timeline graph was no longer drawn. There is another issue though; it only shows the left y-axis, and the packet count value is not scaled properly. This appears to be a bug in vis.js, for which I don't have a solution at the moment.